#0.1.8.3

+++ src/Adapter/Controller.php
@@ -6,7 +6,18 @@ namespace Adapter;
 
 class Controller
 {
-    private $AdapterModels = [];
+    const BOTH  = 0;
+    const LEFT  = 1;
+    const RIGHT = 2;
+    /**
+     * @var bool 简易方法
+     */
+    protected $adapter_function_show = false;
+    protected $adapter_function_view = false;
+    protected $adapter_function_save = false;
+    protected $adapter_function_del  = false;
+    protected $adapter_function_add  = false;
+
     protected $namespace = '';
     protected $AdapterModel;
     protected $output = [];
@@ -18,31 +29,40 @@ class Controller
 
     public function __construct(Model $model = null)
     {
-        $this->param = request()->param();
+        $this->param = empty($this->param)?request()->param():array_merge($this->param,request()->param());
         $this->autoLoadModel($model);
-
-
     }
 
+
+    /**
+     * 自动加载模型
+     * 模型名称与数据表对应
+     * 允许使用前缀
+     * @param Model|null $model 模型对象
+     */
     public function autoLoadModel(Model $model = null){
         self::setModelNamespace();
-
         if(is_subclass_of($model,Model::class)){
+            $model->autoParam($this->param);
             $this->AdapterModel = $model;
-            $this->AdapterModel && $this->AdapterModel->autoParam($this->param);
+            return;
         }else{
-            if (is_string($this->table)){
-                $model_name = self::$modelNamespace.$this->table;
-                $this->AdapterModel = new $model_name();
-            }else{
-                foreach ($this->table as $item){
-                    $this->AdapterModels[$item]=$this->implant(new $item());
-                }
+            $model_class = get_class($this);
+            $model_class = strtolower($model_class);
+            $model_class = str_replace('\\','/',$model_class);
+            $model_class = self::$modelNamespace.basename($model_class);
+
+
+            if(class_exists($model_class)||class_exists($model_class = trim($this->namespace,'\\').'\\'.$this->table,$model_class)){
+
+                $this->AdapterModel = new $model_class();
+                $this->AdapterModel ->autoParam($this->param);
+                return ;
+            }else
+            {
+                throw new AdapterException('不存在数据表模型:'.$model_class);
             }
         }
-
-        if(!$this->AdapterModel)$this->AdapterModel = new Model();
-        $this->AdapterModel->autoParam($this->param);
     }
 
     public function implant(Model $adapterModel = null){
@@ -63,9 +83,10 @@ class Controller
 
 
     public function show(){
-        $this->AdapterModel->autoParam();
+        if($this->adapter_function_show)return Helper::fatal('Invalid access');
+        $this->AdapterModel->autoParam($this->param);
         $this->output = $this->AdapterModel->select();
-//        return $this->output->back();
+
         return Helper::success([
             'page'  =>$this->AdapterModel->page(),
             'limit' =>$this->AdapterModel->limit(),
@@ -76,22 +97,30 @@ class Controller
 
 
     public function view(){
-        $this->AdapterModel->autoParam();
+        if($this->adapter_function_view)return Helper::fatal('Invalid access');
+        $this->AdapterModel->autoParam($this->param);
+
         $this->output = $this->AdapterModel->find();
+
         return Helper::success($this->output);
     }
 
     public function del(){
+        if($this->adapter_function_del)return Helper::fatal('Invalid access');
         return Helper::auto($this->AdapterModel->delete(),[$this->AdapterModel->error()]);
     }
 
     public function add(){
+        if($this->adapter_function_add)return Helper::fatal('Invalid access');
+        $this->AdapterModel->autoParam($this->param);
         return Helper::auto($this->AdapterModel->add(),[$this->AdapterModel->error()]);
     }
 
     public function save(array $extra_condition = []){
-        $this->AdapterModel->autoParam();
-        return Helper::auto($this->AdapterModel->save(),[$this->AdapterModel->error()]);
+        if($this->adapter_function_save)return Helper::fatal('Invalid access');
+        $this->AdapterModel->autoParam($this->param);
+
+        return Helper::auto($this->AdapterModel->save([],$extra_condition),[$this->AdapterModel->error()]);
     }
 
     /**
@@ -128,7 +157,7 @@ class Controller
     }
 
     public function getModel(string $table = ''){
-        return $this->AdapterModel;
+        return $this->AdapterModel->getCursor();
     }
 
     public function choseTable(string $table = null){
@@ -140,14 +169,11 @@ class Controller
             self::$modelNamespace = '\\app\\';
             self::$modelNamespace.= empty($app = app('http')->getName())?'':$app.'\\';
             self::$modelNamespace.= "model\\";
-
             return;
         }else{
             self::$modelNamespace = $namespace;
             return;
         }
-        self::$modelNamespace = '\\app\\model\\';
-        echo self::$modelNamespace;
     }
 
 
@@ -155,18 +181,57 @@ class Controller
      * 必填参数
      */
     public function required(array $required){
-        foreach ($required as $item){
-            if(!isset($this->param[$item])) {
-                $this->error_message = '缺少参数'.(app()->isDebug()?(":".$item):'');
-                return false;
+
+
+        if(empty(array_diff($required,array_intersect($required,array_keys($this->param))))){
+            foreach ($this->param as $key => $item){
+                if(empty($item)){
+                    $this->error_message = $key.' 字段不能为空';
+                    return false;
+                }
             }
+        }else{
+            $this->error_message = '缺少必传字段:'.implode(array_diff($required,array_keys($this->param)),',');
+            return false;
         }
+
         return true;
+
     }
 
     public function error(){
         return $this->error_message;
     }
 
+
+    /**------------------------------------*/
+    /**
+     * @param array $field_name
+     * @param int $model
+     */
+    public function like(array $field_name = [],int $model = self::BOTH){
+        foreach ($this->param as $key=>$item){
+            if(in_array($key,$field_name) && is_string($item)){
+                switch ($model){
+                    case self::BOTH:
+                        $this->param[] = [$key,'LIKE',"%{$item}%"];
+                        break;
+                    case self::LEFT:
+                        $this->param[] = [$key,'LIKE',"%{$item}"];
+                        break;
+                    case self::RIGHT:
+                        $this->param[] = [$key,'LIKE',"{$item}%"];
+                        break;
+                }
+                unset($this->param[$key]);
+            }else{
+                if(is_string($key)){
+                    $this->param[] = [$key,'=',$item];
+                    unset($this->param[$key]);
+                }
+            }
+
+        }
+    }
 
 }
